fix(mrcm): Move, rename attribute type enum
