+++ snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/core/MrcmAttributeType.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2018-2022 B2i Healthcare Pte Ltd, http://b2i.sg
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.b2international.snowowl.snomed.core;
+
+import static com.google.common.base.Strings.nullToEmpty;
+
+import com.b2international.commons.exceptions.BadRequestException;
+
+/**
+ * @since 8.8.0
+ */
+public enum MrcmAttributeType {
+	
+	DATA,
+	OBJECT,
+	ALL;
+	
+	public static MrcmAttributeType getByNameIgnoreCase(String name) {
+		for (final MrcmAttributeType type : values()) {
+			if (nullToEmpty(name).equalsIgnoreCase(type.toString())) {
+				return type;
+			}
+		}
+		throw new BadRequestException("Unknown attribute type: '%s'", name);
+	}
+
+}
